ai-services/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
dnspython==2.7.0
fastapi==0.109.1
h11==0.14.0
idna==3.10
joblib==1.4.2
motor==3.7.0
numpy==2.2.3
pandas==2.2.3
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
starlette==0.35.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.22.0